third_party/blink/web_tests/external/wpt/event-timing/idlharness.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] EventCounts must be primary interface of performance.eventCounts
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Stringification of performance.eventCounts
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] EventCounts interface: performance.eventCounts must not have property "undefined"
  Cannot use 'in' operator to search for 'undefined' in undefined
Harness: the test ran to completion.
